ocm-oidc-idp-api: phase 2a - shared rhidp Vault domain layer
Move SsoClientSecret and cluster_vault_secret_id out of
qontract_api/integrations/sso_client/domain.py into a new shared
qontract_api/rhidp/domain.py, since the not-yet-migrated ocm_oidc_idp
integration needs the exact same Vault secret schema and id function
(byte-compatibility contract established in #5689). Pure move, no
behavior change to sso_client.
